refactor(refactor): :recycle: [sidebar] refactor code
