+++ src/app/[lang]/dashboard/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>Dashboard</div>
+}
