+++ src/app/[lang]/layout.tsx
@@ -1,36 +1,54 @@
-import { Separator } from '@/components/ui/separator'
-import { SidebarInset, SidebarProvider, SidebarTrigger } from '@/components/ui/sidebar'
+import type { Metadata } from 'next'
+import { Be_Vietnam_Pro } from 'next/font/google'
+
+import { SidebarInset, SidebarProvider } from '@/components/ui/sidebar'
 
 import { AppSidebar } from '@/components/app-sidebar'
 import { getDictionary } from '@/dictionaries/get-dictionary'
 import Footer from '@/layouts/footer'
 import Header from '@/layouts/header'
+import { cn } from '@/lib/utils'
+import DictionaryProvider from '@/providers/dictionary-provider'
 import { LocaleEnum } from '@/types/locales'
 
-interface Props {
-  children: React.ReactNode
-  params: { lang: LocaleEnum }
-}
-async function Layout({ children, params }: Props) {
-  const dictionary = await getDictionary(params.lang)
-  return (
-    <>
-      <SidebarProvider>
-        <AppSidebar />
-        <SidebarInset>
-          <header className="flex h-16 shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-12">
-            <div className="flex items-center gap-2 px-4">
-              <SidebarTrigger className="-ml-1" />
-              <Separator orientation="vertical" className="mr-2 h-4" />
-            </div>
-          </header>
-          <div className="flex flex-1 flex-col">{children}</div>
-        </SidebarInset>
-      </SidebarProvider>
-      <Footer dictionary={dictionary} />
-      <Header dictionary={dictionary} />
-    </>
-  )
+import '../globals.scss'
+
+const beVietNamPro = Be_Vietnam_Pro({
+  subsets: ['latin'],
+  weight: ['300', '400', '500', '600', '700'],
+  variable: '--be-vietnam-pro',
+})
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
 }
 
-export default Layout
+export default async function RootLayout({
+  children,
+  params,
+}: Readonly<{
+  children: React.ReactNode
+  params: { lang: LocaleEnum }
+}>) {
+  const dictionary = await getDictionary(params.lang)
+
+  return (
+    <html lang={params.lang}>
+      <body className={cn(beVietNamPro.className, beVietNamPro.variable)}>
+        <DictionaryProvider dictionary={dictionary}>
+          <SidebarProvider>
+            <AppSidebar />
+            <SidebarInset>
+              <Header />
+
+              <div className="flex flex-1 flex-col">{children}</div>
+
+              <Footer dictionary={dictionary} />
+            </SidebarInset>
+          </SidebarProvider>
+        </DictionaryProvider>
+      </body>
+    </html>
+  )
+}
