--- src/app/example-ui/page.tsx
@@ -1,9 +1,0 @@
-import DemoForm from '@/components/form/demo-form'
-
-export default function ExampleUI() {
-  return (
-    <div>
-      <DemoForm />
-    </div>
-  )
-}
