--- src/app/layout.tsx
@@ -1,40 +1,0 @@
-import type { Metadata } from 'next'
-import { Be_Vietnam_Pro } from 'next/font/google'
-
-import { getDictionary } from '@/dictionaries/get-dictionary'
-import { cn } from '@/lib/utils'
-import DictionaryProvider from '@/providers/dictionary-provider'
-import { LocaleEnum } from '@/types/locales'
-
-import './globals.scss'
-
-const beVietNamPro = Be_Vietnam_Pro({
-  subsets: ['latin'],
-  weight: ['300', '400', '500', '600', '700'],
-  variable: '--be-vietnam-pro',
-})
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
-export default async function RootLayout({
-  children,
-  params,
-}: Readonly<{
-  children: React.ReactNode
-  params: { lang: LocaleEnum }
-}>) {
-  const dictionary = await getDictionary(params.lang)
-
-  return (
-    <html lang="en">
-      <body className={cn(beVietNamPro.className, beVietNamPro.variable)}>
-        <DictionaryProvider dictionary={dictionary}>
-          <main className="flex min-h-screen flex-col">{children}</main>
-        </DictionaryProvider>
-      </body>
-    </html>
-  )
-}
