+++ src/layouts/header/index.tsx
@@ -1,11 +1,13 @@
-import { DictionaryProps } from '@/types/common'
+import { Separator } from '@/components/ui/separator'
+import { SidebarTrigger } from '@/components/ui/sidebar'
 
-function Header({ dictionary }: DictionaryProps) {
+function Header() {
   return (
-    <header className="bg-neutral-3 container sticky top-0 z-10 flex items-center justify-between py-4">
-      <span>LOGO</span>
-
-      {dictionary.hello}
+    <header className="flex h-16 shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-12">
+      <div className="flex items-center gap-2 px-4">
+        <SidebarTrigger className="-ml-1" />
+        <Separator orientation="vertical" className="mr-2 h-4" />
+      </div>
     </header>
   )
 }
